Use uint64_t

+++ velox/exec/HashAggregation.cpp
@@ -269,7 +269,8 @@ void HashAggregation::resetPartialOutputIfNeed() {
         "flushRowCount", RuntimeCounter(numOutputRows_));
     lockedStats->addRuntimeStat("flushTimes", RuntimeCounter(1));
     lockedStats->addRuntimeStat(
-        "partialAggregationPct", RuntimeCounter(aggregationPct));
+        "partialAggregationPct",
+        RuntimeCounter(static_cast<uint64_t>(aggregationPct)));
   }
   groupingSet_->resetTable(/*freeTable=*/false);
   partialFull_ = false;
